Draft 1
Initial Draft. Includes Env Var’s for Unisphere server, username, and
password. Script queries for all local VMAX’s talking to Unisphere and
outputs total data sent over SRDFA on a per VMAX/SG basis. This can be
scheduled to be ran once a day to provide historical reporting. It
writes the data out into a csv file where the script is ran from, and
appends all outputs to it.